Convert question from should to must
I've debated this back and forth a few times as the subtly of the
wording change is important.  I believe your question really is along
the must concept though, as it's a natural consequence of the upgrade
path.

I've made this it's own commit though as it is very debatable even in my
own mind and wanted to easily revert it out.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>